api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Stan van Rooy <me@stanvanrooy.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.1.2"
Flask-Cors = "^3.0.10"
PyJWT = "^2.4.0"
Flask-SQLAlchemy = "^2.5.1"

[tool.poetry.dev-dependencies]
pylint = "^2.13.9"
pytest = "^7.1.2"
pytest-watch = "^4.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
